Penambahan fungsi foto profil

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->text('biografi');
             $table->text('alamat');
-            $table->integer('no_telpon');
+            $table->bigInteger('no_telpon');
             $table->string('ProfilePic')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
